requirements.txt
django==4.0.1
djangorestframework==3.13.1
django-rest-passwordreset==1.2.1
django-cors-headers==3.11.0
django-filter==21.1
pillow==9.0.0